rlly bassic refactor
