+++ types.go
@@ -1,0 +1,6 @@
+package main
+
+type Song struct {
+	name string
+	path string
+}
